+++ tests/settings.py
@@ -19,8 +19,10 @@ DATABASES = {
         'PORT': '',
         'OPTIONS': {
             'charset': 'utf8mb4',
-            'init_command': "SET sql_mode='STRICT_TRANS_TABLES', "
-                            "innodb_strict_mode=1",
+            # The most forward compatible sql_mode, for 5.7
+            'init_command': "SET sql_mode='STRICT_TRANS_TABLES,NO_ZERO_DATE,"
+                            "NO_ZERO_IN_DATE,ERROR_FOR_DIVISION_BY_ZERO,"
+                            "NO_AUTO_CREATE_USER', innodb_strict_mode=1",
         },
         'TEST': {
             'COLLATION': "utf8mb4_general_ci",
@@ -36,8 +38,10 @@ DATABASES = {
         'PORT': '',
         'OPTIONS': {
             'charset': 'utf8mb4',
-            'init_command': "SET sql_mode='STRICT_TRANS_TABLES', "
-                            "innodb_strict_mode=1",
+            # The most forward compatible sql_mode, for 5.7
+            'init_command': "SET sql_mode='STRICT_TRANS_TABLES,NO_ZERO_DATE,"
+                            "NO_ZERO_IN_DATE,ERROR_FOR_DIVISION_BY_ZERO,"
+                            "NO_AUTO_CREATE_USER', innodb_strict_mode=1",
         },
         'TEST': {
             'COLLATION': "utf8mb4_general_ci",